[ UPDATE-editpage ] postpage와 컴포넌트를 공용으로 사용하라 - postpage에서 만들어놓은 input / image 컴포넌트를 공용으로 사용하였다 - type:edit을 추가하여 조건부로 edit에 맞는 부분을 렌더링 하였다. - postpage와 다른 점은 editpage에서 mongodb에서 default data를 props로 받아 사용한다는 점이다 - 기존 postdata와 새로 업데이트 된 input / image 데이터를 조합하여 editData를 서버로 요청한다 - 버튼 컴포넌트도 겹치는 부분이 많긴 한데, 조건부로 데이터를 만지기엔 약간 복잡해 질 것 같아 아직 보류중이다.
